+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# x.x.x (Next Release)
+* BUG: Fixed logic on RunAsNonRoot check to incorporate settings in podSpec
+
 # 0.2.0
 * Added `--output-format` flag for better CI/CD support
 * Added `--display-name` flag